Multiple bug fixes.
Fixed an issue where the number of motion controllers was being counted incorrectly leading to a failure to set motion controller movement speed multipliers.

Demoted the severity of printout from error to debug for when the number of motion controllers does not match what is expected from the movement speed multiplier config loader prior to finishing device manager window setup.